requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
redis==5.0.8
celery==5.4.0
pydantic==2.9.2
psycopg2-binary==2.9.9
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
asyncpg==0.30.0
fastapi-cache2==0.2.2
pytest-celery==1.0.0
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==6.1.0
aiosqlite
pytest-asyncio